♻️ REFACTOR: 다이어리 생성, 수정 페이지의 푸터와 로고 컴포넌트화
