+++ src/pages/CreateDiary.jsx
@@ -3,15 +3,13 @@ import { useMutation } from "@tanstack/react-query";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
 import { TiPencil } from "react-icons/ti";
-import coupleDiaryBear from "../assets/images/coupleDiaryBear.webp";
-import soloDiaryBear from "../assets/images/soloDiaryBear.webp";
 import useDispatchHook from "../hooks/useDispatchHook";
 import Diary from "../components/main/Diary/Diary";
 import { mainApi } from "../apis/axios";
 import { flex } from "../UI/common";
 import { Header } from "../components/common/header/Header";
-
-const color = ["#FF8181", "#FFCA7A", "#FFE99A", "#A4F5A3", "#9CDBF7", "#BB9EFA"];
+import ColorPickerFooter from "../components/Create&UpdateDiary/ColorPickerFooter";
+import CreatePageLogo from "../components/Create&UpdateDiary/CreatePageLogo";
 
 const CreateDiary = () => {
   const [selectedColor, setSelectedColor] = useState("");
@@ -45,27 +43,14 @@ const CreateDiary = () => {
         </Header.SpaceBetween>
       </Header>
       <CreateDiaryBox>
-        <CreateLogoBear>
-          <img src={couple === 0 ? soloDiaryBear : coupleDiaryBear} alt="다이어리 생성 곰돌이 그림" />
-          <span>{couple === 0 ? "혼자써요 !" : "같이써요 !"}</span>
-        </CreateLogoBear>
+        <CreatePageLogo />
         <div className="pencilIcon-box">
           <TiPencil />
         </div>
         <input type="text" ref={diaryTitleInputRef} />
         <Diary bgColor={selectedColor} />
       </CreateDiaryBox>
-      <Footer>
-        {color.map((color, i) => {
-          return (
-            <ColorPicker
-              key={`diaryColorPicker${i}`}
-              color={color}
-              onClick={() => setSelectedColor(color)}
-            ></ColorPicker>
-          );
-        })}
-      </Footer>
+      <ColorPickerFooter setSelectedColor={setSelectedColor} />
     </>
   );
 };
@@ -89,47 +74,5 @@ export const CreateDiaryBox = styled.section`
     position: absolute;
     top: calc(50% - 18.75rem);
     left: calc(50% + 7.5rem);
-  }
-`;
-
-export const CreateLogoBear = styled.div`
-  ${flex("", "", "", "column")}
-  position: absolute;
-  top: 10%;
-  left: calc(50% - 15rem);
-  cursor: pointer;
-  img {
-    width: 4.2rem;
-    height: 4.2rem;
-  }
-  span {
-    font-size: 1rem;
-    margin-top: 0.6rem;
-  }
-`;
-
-export const Footer = styled.footer`
-  position: absolute;
-  bottom: 2%;
-  left: 0;
-  width: 100%;
-  height: 7.2rem;
-  background-color: white;
-  ${flex("space-evenly", "")}
-`;
-
-export const ColorPicker = styled.button`
-  width: 3.6rem;
-  height: 3.6rem;
-  border-radius: 4px;
-  border: none;
-  cursor: pointer;
-  background-color: ${(props) => props.color};
-  transition: all 0.3s;
-  :hover {
-    transform: scale(1.1);
-  }
-  :focus {
-    transform: scale(1.1);
   }
 `;
